Refactored Profile page
